Don't use absolute paths for redirects.
So that we can do simple redirects from the webserver.

Right now, for bot hosted on subpath, we're using the webserver to
rewrite requests when passing it to the reverse proxy. we can do the
same for the respose also, but this seemed better.

Instead of rewriting at all, we can use a Router and use that to move
all endpoints to a specific subpath, but then the config gets muddy.

Going with this, for now.