Use webhooks for updates
As casbot is running as a systemctl service, we have to restart casbot
after git pulling. And, as centos doesn't support user units, we have to
to use sudo. **Which is not ideal**. but shouldn't be a problem, as we
are only allowing it to execute the specific restart command in sudoers.